Add cardinality note to Quad Sections and Coverage
Discovered by irrelevant.irreverent. Behavior is as follows:

If faces with a normal facing in the -Y or +-X direction are assigned to the North section for culling, culling will be broken ingame. The faces will appear culled even without a brick next to it to cull the face. Cullable faces are required to point towards one of the cardinal directions their normal corresponds to, with +Y as north, +X as east, -Y as south, -X as west.